+++ core/src/main/java/org/twins/core/dao/draft/DraftTwinFieldDataListRepository.java
@@ -1,12 +1,12 @@
 package org.twins.core.dao.draft;
 
-import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
 
@@ -33,7 +33,7 @@ public interface DraftTwinFieldDataListRepository extends CrudRepository<DraftTw
                     "from draft_twin_field_data_list " +
                     "where draft_id = :draftId " +
                     "  and cud_id = 'CREATE';")
-    long commitCreates(@Param("draftId") UUID draftId);
+    int commitCreates(@Param("draftId") UUID draftId);
 
     @Transactional
     @Modifying
@@ -44,7 +44,7 @@ public interface DraftTwinFieldDataListRepository extends CrudRepository<DraftTw
                     "where draft_id = :draftId " +
                     "  and dta.twin_field_data_list_id = twin_field_data_list.id " +
                     "  and dta.cud_id = 'UPDATE';")
-    long commitUpdates(@Param("draftId") UUID id);
+    int commitUpdates(@Param("draftId") UUID id);
 
     @Transactional
     @Modifying
@@ -53,5 +53,5 @@ public interface DraftTwinFieldDataListRepository extends CrudRepository<DraftTw
                     "using draft_twin_field_data_list dta " +
                     "where dta.draft_id = :draftId " +
                     "and dta.twin_field_data_list_id = ta.id and cud_id = 'DELETE'")
-    long commitDeletes(@Param("draftId") UUID id);
+    int commitDeletes(@Param("draftId") UUID id);
 }
